Delete deprecated Today, update DaylightWidget
