--- Today/UIColor+Helpers.swift
@@ -1,33 +1,0 @@
-//
-//  UIColor+Helpers.swift
-//  Project
-//
-//  Created by Dylan Marriott on 19/02/17.
-//
-
-import Foundation
-import UIKit
-
-// http://stackoverflow.com/a/38435309/1472887
-extension UIColor {
-
-    func lighter(by percentage: CGFloat = 30.0) -> UIColor {
-        return self.adjust(by: abs(percentage))
-    }
-
-    func darker(by percentage: CGFloat = 30.0) -> UIColor {
-        return self.adjust(by: -1 * abs(percentage))
-    }
-
-    func adjust(by percentage: CGFloat = 30.0) -> UIColor {
-        var r: CGFloat = 0, g: CGFloat = 0, b: CGFloat = 0, a: CGFloat = 0
-        if self.getRed(&r, green: &g, blue: &b, alpha: &a) {
-            return UIColor(red: min(r + percentage / 100, 1.0),
-                           green: min(g + percentage / 100, 1.0),
-                           blue: min(b + percentage / 100, 1.0),
-                           alpha: a)
-        } else {
-            return self
-        }
-    }
-}
